+++ pkg/state/create.go
@@ -223,6 +223,10 @@ func (c *StateCreator) ParseAndLoad(content []byte, baseDir, file string, envNam
 	}
 	state.RenderedValues = vals
 
+	for _, w := range state.reservedLabelWarnings() {
+		c.logger.Warnf("WARNING: %s", w)
+	}
+
 	return state, nil
 }